tinyworld-backend/pyproject.toml
[project]
name = "tinyworld-api"
version = "0.1.0"
description = "A tiny autonomous world simulation backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "websockets>=12.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-mongodb>=0.4.0",
    "langgraph>=0.2.70",
    "langgraph-checkpoint-mongodb>=0.1.0",
    "opik>=1.4.11",
    "motor>=3.5.0",
    "pymongo>=4.8.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
    "langchain-chroma>=0.1.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tinyworld"]
